# docs/c4-membership-onboarding/api/openapi.fragment.yaml
openapi: 3.1.0
info:
  title: FEELM C4A Local Email Membership and Onboarding API
  version: 0.2.0-c4a-local
  summary: local 이메일 membership, LIKE/DISLIKE onboarding, KR OTT subscription 계약
  description: |
    APPROVED_LOCAL_PROFILE_WITH_BLOCKED_PRODUCTION_EXTENSIONS. 보수 token은 local profile에만
    채택됐다. main OpenAPI, production email, OAuth, restart, password/account lifecycle은 차단한다.
servers:
  - url: http://localhost:8080
security:
  - bearerAuth: []
...
paths:
  /api/v1/auth/sign-up:
    post:
      tags: [Membership]
      operationId: createEmailSignup
      x-implementation-status: AUTHORIZED_LOCAL_CONTRACT
      summary: 이메일·비밀번호·nickname으로 pending membership 생성
      security: []
      x-local-decision-authority: [DN-C4A-002, DN-C4A-003]
      x-enumeration-response-policy:
        status: AUTHORIZED_LOCAL_PROFILE
        actualHandle: persistent opaque UUID for the new REAL public flow, distinct from every challenge ID
        duplicateOrUnknownHandle: persistent decoy opaque UUID with no account, secret, or mail
        returnedHandleRetentionHours: 24
        followUpState: same persisted cooldown, attempts, expiry, and revision for actual and decoy
        publicShape: masked email, QUEUED, response-time plus 10-minute expiry, response-time plus 60-second resend, revision 1
        responseFloor: calibrated commonFloor=max(p99 supported PHC workloads)+25ms
        responseJitterMilliseconds: {minimum: 0, maximum: 75}
        duplicateWork: dummy Argon2id plus queue-equivalent work
        stableHandleEntity: EMAIL_SIGNUP_PUBLIC_FLOW
        currentChallenge: versioned internal current_challenge_id for REAL only
        flowExpires: createdAt plus 24 hours without resend extension
        pendingRecovery: only after prior REAL flow is EXPIRED and before pendingPurgeAt; preserve nickname and OTT set
        pendingPurge: account createdAt plus 30 days without re-signup extension
        exactPublicCodes:
          malformed: {status: 400, code: VALIDATION_ERROR}
          idempotencyConflict: {status: 409, code: IDEMPOTENCY_KEY_REUSED}
          stateConflict: {status: 409, code: AUTH_STATE_CONFLICT}
          throttled: {status: 429, code: AUTH_FLOW_THROTTLED}
          dependency: {status: 503, code: AUTH_DEPENDENCY_UNAVAILABLE}
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateEmailSignupRequest'
      responses:
        '202':
          description: >-
            존재 여부와 무관하게 같은 shape인 generic pending 상태. signupId와 expiry/resend 시각은
            신규 account에는 actual challenge handle이고 duplicate/unknown에는 persisted decoy handle이다.
            raw verification secret을 포함하지 않는다.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PendingEmailSignup'
        '400': {$ref: '#/components/responses/PublicSignupBadRequest'}
        '409': {$ref: '#/components/responses/PublicAuthConflict'}
        '429': {$ref: '#/components/responses/AuthThrottled'}
        '503': {$ref: '#/components/responses/AuthDependencyUnavailable'}
  /api/v1/auth/email-verifications:
    post:
      tags: [Membership]
      operationId: verifySignupEmail
      x-implementation-status: AUTHORIZED_LOCAL_CONTRACT
      summary: email verification challenge 소비와 membership 활성화
      security: []
      x-local-decision-authority: [DN-C4A-003]
      x-generic-handle-policy:
        persistence: returned actual and decoy rows exist for 24 hours
        decoyBehavior: every submitted secret follows the wrong-attempt transition
        concurrency: row lock gives verify or resend exactly one transition winner
        unknownHandle: generic invalid-or-expired response without account detail
        exactPublicCodes:
          malformed: {status: 400, code: VALIDATION_ERROR}
          wrong: {status: 400, code: VERIFICATION_INVALID}
          unknownTerminalOrExpired: {status: 400, code: SIGNUP_FLOW_INVALID_OR_EXPIRED}
          attemptsExhausted: {status: 429, code: VERIFICATION_ATTEMPTS_EXHAUSTED}
          throttled: {status: 429, code: AUTH_FLOW_THROTTLED}
          idempotencyConflict: {status: 409, code: IDEMPOTENCY_KEY_REUSED}
          stateConflict: {status: 409, code: AUTH_STATE_CONFLICT}
          dependency: {status: 503, code: AUTH_DEPENDENCY_UNAVAILABLE}
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VerifyEmailRequest'
      responses:
        '200':
          description: membership ACTIVE. 자동 login/token 발급 없이 login 화면으로 진행한다.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EmailVerificationResult'
        '400': {$ref: '#/components/responses/PublicVerificationBadRequest'}
        '409': {$ref: '#/components/responses/PublicAuthConflict'}
        '429': {$ref: '#/components/responses/AuthFlowRateLimited'}
        '503': {$ref: '#/components/responses/AuthDependencyUnavailable'}
  /api/v1/auth/email-verification-resends:
    post:
      tags: [Membership]
      operationId: resendSignupEmailVerification
      x-implementation-status: AUTHORIZED_LOCAL_CONTRACT
      summary: pending signup email verification 재전송 요청
      security: []
      x-local-decision-authority: [DN-C4A-003]
      x-generic-handle-policy:
        persistence: returned actual and decoy rows exist for 24 hours
        decoyBehavior: update the same cooldown, attempt, expiry, and revision state without mail
        concurrency: row lock gives verify or resend exactly one transition winner
        unknownHandle: generic invalid-or-expired response without account detail
        stableSignupId: unchanged; REAL current_challenge_id/version and flow revision advance
        exactPublicCodes:
          accepted: {status: 202}
          invalidOrExpired: {status: 400, code: SIGNUP_FLOW_INVALID_OR_EXPIRED}
          throttled: {status: 429, code: AUTH_FLOW_THROTTLED}
          idempotencyConflict: {status: 409, code: IDEMPOTENCY_KEY_REUSED}
          stateConflict: {status: 409, code: AUTH_STATE_CONFLICT}
          dependency: {status: 503, code: AUTH_DEPENDENCY_UNAVAILABLE}
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ResendEmailVerificationRequest'
      responses:
        '202':
          description: 열거를 줄이는 generic delivery state. raw secret 없음.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VerificationDeliveryState'
        '400': {$ref: '#/components/responses/PublicResendBadRequest'}
        '409': {$ref: '#/components/responses/PublicAuthConflict'}
        '429': {$ref: '#/components/responses/AuthThrottled'}
        '503': {$ref: '#/components/responses/AuthDependencyUnavailable'}
  /api/v1/auth/login:
    post:
      tags: [Authentication]
      operationId: loginWithEmail
      x-implementation-status: AUTHORIZED_LOCAL_CONTRACT
      summary: ACTIVE email membership 로그인
      security: []
      x-local-decision-authority: [DN-C4A-001, DN-C4A-003]
      x-enumeration-timing-policy:
        exactPublicCodes:
          malformed: {status: 400, code: VALIDATION_ERROR}
          invalidCredentials: {status: 401, code: INVALID_CREDENTIALS}
          originFailure: {status: 403, code: AUTH_ORIGIN_FORBIDDEN}
          pendingAfterCorrectPassword: {status: 403, code: EMAIL_VERIFICATION_REQUIRED}
          throttled: {status: 429, code: AUTH_FLOW_THROTTLED}
          dependency: {status: 503, code: AUTH_DEPENDENCY_UNAVAILABLE}
        originFailure: {status: 403, code: AUTH_ORIGIN_FORBIDDEN, beforeCredentialWork: true}
        unknownEmail: verify fixed dummy PHC hash with current production Argon2id parameters
        wrongPassword: verify stored PHC hash
        commonFailure: {status: 401, code: INVALID_CREDENTIALS, timing: 'calibrated commonFloor=max(p99 supported PHC workloads)+25ms plus CSPRNG 0..75ms jitter'}
        pendingAfterCorrectPassword: {status: 403, code: EMAIL_VERIFICATION_REQUIRED}
      parameters:
        - $ref: '#/components/parameters/OriginHeader'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EmailLoginRequest'
      responses:
        '200':
          description: DN-C4A-001의 proposed access bearer와 선택 cookie profile의 refresh/CSRF cookie 발급
          x-set-cookie-contract:
            repeatedHeader: Set-Cookie
            count: 2
            commaFolding: forbidden
            action: issue
            profiles: [productionHttps, localHttpLoopbackOnly]
          headers:
            Set-Cookie:
              description: >-
                production HTTPS profile은 __Host-feelm_refresh (Secure, HttpOnly, SameSite=Lax, Path=/)와
                JavaScript-readable __Host-feelm_csrf (Secure, SameSite=Lax, Path=/)를 발급한다.
                loopback local HTTP profile은 feelm_local_refresh (non-Secure, HttpOnly, SameSite=Lax, Path=/)와
                JavaScript-readable feelm_local_csrf (non-Secure, SameSite=Lax, Path=/)를 발급한다.
                선택 profile의 두 header를 각각 전송하며 Domain은 없고 profile을 섞지 않는다.
              schema: {type: string}
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthenticationResult'
        '400': {$ref: '#/components/responses/LoginBadRequest'}
        '401': {$ref: '#/components/responses/InvalidCredentials'}
        '403': {$ref: '#/components/responses/LoginForbidden'}
        '429': {$ref: '#/components/responses/AuthThrottled'}
        '503': {$ref: '#/components/responses/AuthDependencyUnavailable'}
  /api/v1/auth/refresh:
    post:
      tags: [Authentication]
      operationId: refreshAuthentication
      x-implementation-status: AUTHORIZED_LOCAL_CONTRACT
      summary: proposed refresh rotation으로 access bearer 갱신
      security:
        - refreshCookie: []
          csrfCookie: []
        - localRefreshCookie: []
          localCsrfCookie: []
      x-local-decision-authority: [DN-C4A-001]
      x-race-clock: PostgreSQL primary clock_timestamp under session and token row lock
      x-cookie-profile-auth: exactly one complete production or local pair; mixed or partial pairs are forbidden
      x-exact-errors:
        invalidSession: {status: 401, code: AUTH_SESSION_INVALID}
        origin: {status: 403, code: AUTH_ORIGIN_FORBIDDEN}
        csrf: {status: 403, code: CSRF_FORBIDDEN}
        race: {status: 409, code: REFRESH_RACE_RETRY_NEW_COOKIE}
        throttled: {status: 429, code: AUTH_FLOW_THROTTLED}
        dependency: {status: 503, code: AUTH_DEPENDENCY_UNAVAILABLE}
      parameters:
        - $ref: '#/components/parameters/OriginHeader'
        - $ref: '#/components/parameters/CsrfTokenHeader'
      responses:
        '200':
          description: 회전된 refresh cookie와 새 access bearer
          x-set-cookie-contract:
            repeatedHeader: Set-Cookie
            count: 2
            commaFolding: forbidden
            action: rotate
            profiles: [productionHttps, localHttpLoopbackOnly]
          headers:
            Set-Cookie:
              description: >-
                새 refresh와 새 signed CSRF cookie를 별도 Set-Cookie header로 함께 회전한다.
                이전 CSRF header/cookie pair는 새 refresh generation에 사용할 수 없다.
              schema: {type: string}
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthenticationResult'
        '401': {$ref: '#/components/responses/AuthSessionInvalid'}
        '403': {$ref: '#/components/responses/AuthMutationForbidden'}
        '409': {$ref: '#/components/responses/RefreshRaceConflict'}
        '429': {$ref: '#/components/responses/AuthThrottled'}
        '503': {$ref: '#/components/responses/AuthDependencyUnavailable'}
  /api/v1/auth/logout:
    post:
      tags: [Authentication]
      operationId: logoutCurrentSession
      x-implementation-status: AUTHORIZED_LOCAL_CONTRACT
      summary: refresh cookie가 가리키는 current authentication session 폐기
      security:
        - refreshCookie: []
          csrfCookie: []
        - localRefreshCookie: []
          localCsrfCookie: []
        - {}
      x-local-decision-authority: [DN-C4A-001]
      x-logout-retry-policy:
        origin: exact allowed Origin is mandatory for every branch, including cookie-less retry
        bothAuthCookiesAbsent: 204 with both exact clear headers; no CSRF, idempotency, DB, or audit mutation required
        partialPair: 403 CSRF_FORBIDDEN
        completePairInvalidSession: valid Origin and CSRF but unknown, expired, or revoked refresh returns 401 AUTH_SESSION_INVALID
        successfulReplay: same idempotency key replays 204 and exact clear without another audit mutation
        clearAttributes: selected profile issuance attributes plus Max-Age=0 and epoch Expires
        exactErrors:
          origin: {status: 403, code: AUTH_ORIGIN_FORBIDDEN}
          csrfOrPartialPair: {status: 403, code: CSRF_FORBIDDEN}
          invalidSession: {status: 401, code: AUTH_SESSION_INVALID}
          idempotencyConflict: {status: 409, code: IDEMPOTENCY_KEY_REUSED}
          dependency: {status: 503, code: AUTH_DEPENDENCY_UNAVAILABLE}
      parameters:
        - $ref: '#/components/parameters/OptionalIdempotencyKey'
        - $ref: '#/components/parameters/OriginHeader'
        - $ref: '#/components/parameters/OptionalCsrfTokenHeader'
      responses:
        '204':
          description: >-
            current refresh family를 즉시 revoke하고 refresh/CSRF cookie를 제거한다. 이미 발급된
            access JWT는 server-side denylist가 없는 v1 후보에서 최대 10분 exp까지 유효할 수 있다.
          x-set-cookie-contract:
            repeatedHeader: Set-Cookie
            count: 2
            commaFolding: forbidden
            action: clear
            profiles: [productionHttps, localHttpLoopbackOnly]
          headers:
            Set-Cookie:
              description: >-
                선택 profile의 refresh/CSRF name, Path=/, no Domain, SameSite=Lax, Secure/HttpOnly를 발급과
                동일하게 유지하고 Max-Age=0, Expires=Thu, 01 Jan 1970 00:00:00 GMT를 더한 두 header.
              schema: {type: string}
        '401': {$ref: '#/components/responses/AuthSessionInvalid'}
        '403': {$ref: '#/components/responses/AuthMutationForbidden'}
        '409': {$ref: '#/components/responses/IdempotencyConflict'}
        '503': {$ref: '#/components/responses/AuthDependencyUnavailable'}
  /api/v1/me:
    get:
      tags: [MyProfile]
      operationId: getMyMembership
      x-implementation-status: AUTHORIZED_LOCAL_CONTRACT
      summary: 내 membership·nickname·onboarding 요약
      x-local-decision-authority: [DN-C4A-001]
      security: [{bearerAuth: []}]
      responses:
        '200':
          description: internal userId와 raw email을 포함하지 않는 내 요약
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MyMembership'
        '401': {$ref: '#/components/responses/Unauthorized'}
        '503': {$ref: '#/components/responses/ServiceUnavailable'}
    patch:
      tags: [MyProfile]
      operationId: updateMyNickname
      x-implementation-status: AUTHORIZED_LOCAL_CONTRACT
      summary: 내 nickname 변경
      x-local-decision-authority: [DN-C4A-001, DN-C4A-002]
      security: [{bearerAuth: []}]
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
        - $ref: '#/components/parameters/ExpectedRevision'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateNicknameRequest'
      responses:
        '200':
          description: nickname과 profile revision 갱신
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MyMembership'
        '400': {$ref: '#/components/responses/ValidationError'}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '409': {$ref: '#/components/responses/Conflict'}
        '503': {$ref: '#/components/responses/ServiceUnavailable'}
  /api/v1/onboarding/movies:
    get:
      tags: [Onboarding]
      operationId: listOnboardingMovies
      x-implementation-status: AUTHORIZED_LOCAL_CONTRACT
      summary: versioned 초기 LIKE/DISLIKE movie 목록
      x-local-decision-authority: [DN-C4A-001, DN-C4A-004]
      security: [{bearerAuth: []}]
      responses:
        '200':
          description: 후보가 없어도 200과 빈 items. local versioned baseline이며 개인화·production 성능 주장이 아니다.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OnboardingMoviePage'
        '401': {$ref: '#/components/responses/Unauthorized'}
        '503': {$ref: '#/components/responses/ServiceUnavailable'}
  /api/v1/onboarding/preferences:
    put:
      tags: [Onboarding]
      operationId: replaceOnboardingPreferences
      x-implementation-status: AUTHORIZED_LOCAL_CONTRACT
      summary: 현재 journey LIKE/DISLIKE 전체 set 교체
      x-local-decision-authority: [DN-C4A-001, DN-C4A-004]
      security: [{bearerAuth: []}]
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
        - $ref: '#/components/parameters/ExpectedRevision'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReplaceOnboardingPreferencesRequest'
      responses:
        '200':
          description: Rating과 분리된 preference set 저장
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OnboardingState'
        '400': {$ref: '#/components/responses/ValidationError'}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '409': {$ref: '#/components/responses/Conflict'}
        '503': {$ref: '#/components/responses/ServiceUnavailable'}
  /api/v1/onboarding/complete:
    post:
      tags: [Onboarding]
      operationId: completeOnboarding
      x-implementation-status: AUTHORIZED_LOCAL_CONTRACT
      summary: SUBMITTED 또는 SKIPPED로 current journey 완료
      x-local-decision-authority: [DN-C4A-001, DN-C4A-004]
      security: [{bearerAuth: []}]
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
        - $ref: '#/components/parameters/ExpectedRevision'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CompleteOnboardingRequest'
      responses:
        '200':
          description: 추천 계산 완료가 아닌 onboarding state 완료
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OnboardingState'
        '400': {$ref: '#/components/responses/ValidationError'}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '409': {$ref: '#/components/responses/Conflict'}
        '503': {$ref: '#/components/responses/ServiceUnavailable'}
...
  /api/v1/me/ott-subscriptions:
    get:
      tags: [MyOttSubscriptions]
      operationId: getMyOttSubscriptions
      x-implementation-status: AUTHORIZED_LOCAL_CONTRACT
      summary: 내 KR OTT subscription set 상태
      x-local-decision-authority: [DN-C4A-001]
      security: [{bearerAuth: []}]
      responses:
        '200':
          description: NOT_CONFIGURED/CONFIGURED/SKIPPED와 provider set
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MyOttSubscriptionSet'
        '401': {$ref: '#/components/responses/Unauthorized'}
        '503': {$ref: '#/components/responses/ServiceUnavailable'}
    put:
      tags: [MyOttSubscriptions]
      operationId: replaceMyOttSubscriptions
      x-implementation-status: AUTHORIZED_LOCAL_CONTRACT
      summary: 내 KR OTT subscription set 전체 교체 또는 skip
      x-local-decision-authority: [DN-C4A-001]
      security: [{bearerAuth: []}]
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
        - $ref: '#/components/parameters/ExpectedRevision'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReplaceOttSubscriptionsRequest'
      responses:
        '200':
          description: atomic set replacement
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MyOttSubscriptionSet'
        '400': {$ref: '#/components/responses/ValidationError'}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '404': {$ref: '#/components/responses/ResourceNotFound'}
        '409': {$ref: '#/components/responses/Conflict'}
        '503': {$ref: '#/components/responses/ServiceUnavailable'}
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: DN-C4A-001 proposed; 현재 C1 required bearer 호환안
    refreshCookie:
      type: apiKey
      in: cookie
      name: __Host-feelm_refresh
      description: Secure, HttpOnly, SameSite=Lax, Path=/, no Domain인 rotating opaque refresh 후보
    csrfCookie:
      type: apiKey
      in: cookie
      name: __Host-feelm_csrf
      description: >-
        Secure, SameSite=Lax, Path=/, no Domain. refresh generation/session에 서명·결합되며
        JavaScript가 같은 값을 X-CSRF-Token으로 제출한다.
    localRefreshCookie:
      type: apiKey
      in: cookie
      name: feelm_local_refresh
      description: loopback local HTTP 전용 non-Secure, HttpOnly, SameSite=Lax, Path=/, no Domain refresh cookie
    localCsrfCookie:
      type: apiKey
      in: cookie
      name: feelm_local_csrf
      description: loopback local HTTP 전용 non-Secure, JavaScript-readable, SameSite=Lax, Path=/, no Domain CSRF cookie
  parameters:
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      schema:
        type: string
        minLength: 8
        maxLength: 128
        pattern: '^[!-~]+$'
...
    OriginHeader:
      name: Origin
      in: header
      required: true
      description: login/refresh/logout은 server exact allowlist와 일치해야 하며 absent/null도 mutation 전에 거부한다.
      schema:
        type: string
        format: uri
...
  responses:
...
    PublicSignupBadRequest:
      description: signup validation failure
      content:
        application/json:
          schema: {$ref: '#/components/schemas/PublicSignupBadRequestError'}
    PublicVerificationBadRequest:
      description: exact verification invalid or expired state
      content:
        application/json:
          schema: {$ref: '#/components/schemas/PublicVerificationBadRequestError'}
    PublicResendBadRequest:
      description: exact resend invalid or expired state
      content:
        application/json:
          schema: {$ref: '#/components/schemas/PublicResendBadRequestError'}
    LoginBadRequest:
      description: exact login request validation failure
      content:
        application/json:
          schema: {$ref: '#/components/schemas/LoginBadRequestError'}
    PublicAuthConflict:
      description: exact public-auth idempotency or state conflict
      content:
        application/json:
          schema: {$ref: '#/components/schemas/PublicAuthConflictError'}
    AuthFlowRateLimited:
      description: exact public-auth throttle or verification-attempt exhaustion
      headers:
        Retry-After:
          schema: {type: integer, minimum: 1}
      content:
        application/json:
          schema: {$ref: '#/components/schemas/AuthFlowRateLimitedError'}
    AuthThrottled:
      description: exact distributed auth throttle
      headers:
        Retry-After:
          schema: {type: integer, minimum: 1}
      content:
        application/json:
          schema: {$ref: '#/components/schemas/AuthThrottledError'}
    AuthDependencyUnavailable:
      description: public auth authority unavailable before mutation
      content:
        application/json:
          schema: {$ref: '#/components/schemas/AuthDependencyUnavailableError'}
...
  schemas:
    CreateEmailSignupRequest:
      type: object
      additionalProperties: false
      required: [email, password, nickname]
      x-local-decision-authority: [DN-C4A-002, DN-C4A-003]
      properties:
        email:
          type: string
          format: email
          maxLength: 320
        password:
          type: string
          format: password
          minLength: 15
          maxLength: 128
          description: DN-C4A-002 packet의 proposed single-factor password 15~128자
        nickname:
          type: string
          minLength: 2
          maxLength: 20
          description: trim+NFKC+casefold와 Letter/Number/underscore 전역 unique proposed
    PendingEmailSignup:
      type: object
      additionalProperties: false
      required: [signupId, membershipStatus, emailMasked, deliveryStatus, verificationExpiresAt, resendAvailableAt, revision]
      properties:
        signupId: {type: string, format: uuid}
        membershipStatus:
          type: string
          const: PENDING_EMAIL_VERIFICATION
        emailMasked: {type: string}
        deliveryStatus:
          type: string
          const: QUEUED
          description: 실제 delivery/account 존재 상태를 노출하지 않는 public generic 값
        verificationExpiresAt:
          type: string
          format: date-time
          description: actual 또는 persisted decoy signup handle 모두 response 기준 now+10m 후보
        resendAvailableAt:
          type: string
          format: date-time
          description: actual 또는 persisted decoy signup handle 모두 response 기준 now+60s 후보
        revision: {type: integer, const: 1}
    VerifyEmailRequest:
      type: object
      additionalProperties: false
      required: [signupId, verificationSecret]
      properties:
        signupId: {type: string, format: uuid}
        verificationSecret:
          type: string
          minLength: 1
          maxLength: 256
          writeOnly: true
          description: >-
            request Host/Forwarded/X-Forwarded-Host가 아니라 pinned server config의 exact HTTPS frontend
            origin fragment로만 전달한다. frontend가 즉시 읽고 history.replaceState로 제거한 뒤 이 POST
            body에만 전송한다. raw 저장·로그·응답 금지.
    ResendEmailVerificationRequest:
      type: object
      additionalProperties: false
      required: [signupId]
      properties:
        signupId: {type: string, format: uuid}
    VerificationDeliveryState:
      type: object
      additionalProperties: false
      required: [signupId, deliveryStatus, verificationExpiresAt, resendAvailableAt, revision]
      properties:
        signupId:
          type: string
          format: uuid
          description: request handle echo; actual/decoy 여부를 나타내지 않음
        deliveryStatus:
          type: string
          const: QUEUED
        verificationExpiresAt: {type: string, format: date-time}
        resendAvailableAt: {type: string, format: date-time}
        revision: {type: integer, minimum: 2, description: stable signupId public-flow revision}
    EmailVerificationResult:
      type: object
      additionalProperties: false
      required: [membershipStatus, emailMasked, nextAction, revision]
      properties:
        membershipStatus: {type: string, const: ACTIVE}
        emailMasked: {type: string}
        nextAction: {type: string, const: LOGIN}
        revision: {type: integer, minimum: 1}
    EmailLoginRequest:
      type: object
      additionalProperties: false
      required: [email, password]
      properties:
        email: {type: string, format: email, maxLength: 320}
        password:
          type: string
          format: password
          minLength: 15
          maxLength: 128
          writeOnly: true
...